% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessoryFuncs.R
\name{darken}
\alias{darken}
\title{Method for plotting.}
\usage{
darken(colour, factor = 1.4)
}
\arguments{
\item{colour}{hex xolor code. Character.}

\item{factor}{factor by which to darken colour. Numeric.}
}
\description{
Method for plotting.
}
\examples{
darken("#2E358B", factor=1.4)
}
\keyword{internal}
