% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileSetup.R
\name{fileSetup}
\alias{fileSetup}
\title{Setup directory for project}
\usage{
fileSetup(directory, date = T)
}
\arguments{
\item{directory}{Absolute path to main/parent diectory. Character.}

\item{date}{Whether to include the date in parent directory name. Logical.}
}
\description{
This function takes output from jailbreakr for the WASH tab on the SDP templates and returns a long format table.
}
\examples{
fileSetup(~/results)
}
\keyword{setup}
