% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatEducationDept.R
\name{formatEducationDept}
\alias{formatEducationDept}
\title{Format EDUCATION tab}
\usage{
formatEducationDept(dat)
}
\arguments{
\item{dat}{List of tiny little matrices extracted from sector templates with jailbreakr}
}
\description{
This function takes output from jailbreakr for the EDUCATION tab on the SDP templates and returns a long format table.
}
\examples{
formatEducationDept(education.raw)
}
\keyword{names}
