% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatFamilyDept.R
\name{formatFamilyDept}
\alias{formatFamilyDept}
\title{Format FAMILY PLANNING  tab}
\usage{
formatFamilyDept(df)
}
\arguments{
\item{dat}{List of tiny little matrices extracted from sector templates with jailbreakr}
}
\description{
This function takes output from jailbreakr for the FAMILYPLANNING tab on the SDP templates and returns a long format table.
}
\examples{
formatFamilyDept(family.raw)
}
\keyword{names}
