% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatHumanitarianDept.R
\name{formatHumanitarianDept}
\alias{formatHumanitarianDept}
\title{Format HUMANITARIAN  tab}
\usage{
formatHumanitarianDept(df)
}
\arguments{
\item{dat}{List of tiny little matrices extracted from sector templates with jailbreakr}
}
\description{
This function takes output from jailbreakr for the HUMANITARIAN tab on the SDP templates and returns a long format table.
}
\examples{
formatHumanitarianDept(humanitarian.raw)
}
\keyword{names}
