% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatNutritionDept.R
\name{formatNutritionDept}
\alias{formatNutritionDept}
\title{Format NUTRITION  tab}
\usage{
formatNutritionDept(dat)
}
\arguments{
\item{dat}{List of tiny little matrices extracted from sector templates with jailbreakr}
}
\description{
This function takes output from jailbreakr for the NUTRITION tab on the SDP templates and returns a long format table.
}
\examples{
formatNutritionDept(nutrition.raw)
}
\keyword{names}
