% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatWashDept.R
\name{formatWashDept}
\alias{formatWashDept}
\title{Format WASH tab}
\usage{
formatWashDept(dat)
}
\arguments{
\item{dat}{List of tiny little matrices extracted from sector templates with jailbreakr}
}
\description{
This function takes output from jailbreakr for the WASH tab on the SDP templates and returns a long format table.
}
\examples{
formatWashDept(wash.raw)
}
\keyword{names}
