% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessoryFuncs.R
\name{getFragStates}
\alias{getFragStates}
\title{Method for lookup table.}
\usage{
getFragStates(lookup, dept_pub_names = T)
}
\arguments{
\item{lookup}{a lookup table.}

\item{dept_pub_names}{use publication names rather than dataset names. Logical.}
}
\description{
Method for lookup table.
}
\examples{
getFragStates(lookup)
}
\keyword{internal}
