% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessoryFuncs.R
\name{getPubNames}
\alias{getPubNames}
\title{Accessory function for Results. Mainly methods for doing lookups.}
\usage{
getPubNames(lookup)
}
\arguments{
\item{lookup}{a lookup table.}

\item{dept_pub_names}{use publication names rather than dataset names. Logical.}
}
\description{
Method for lookup table.
}
\examples{
getPubNames(lookup)
}
\keyword{internal}
